+++ react/inputs/StyleguideInput/GenderInput.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types'
 import { intlShape, injectIntl } from 'react-intl'
 import Dropdown from '@vtex/styleguide/lib/Dropdown'
 import AutocompleteInput from './AutocompleteInput'
-import RuleFieldShape from '../../propTypes/RuleFieldShape'
-import ProfileFieldShape from '../../propTypes/ProfileFieldShape'
-import genders from '../../data/genders'
+import RuleFieldShape from '../../RuleFieldShape'
+import ProfileFieldShape from '../../ProfileFieldShape'
+import genders from '../../modules/genders'
 
 class GenderInput extends Component {
   constructor(props) {