+++ react/modules/validateProfile.js
@@ -1,31 +1,3 @@
-export function addValidation(profile, rules) {
-  const allRules = [...rules.personalFields, ...rules.businessFields]
-
-  return allRules
-    .map(field => ({
-      [field.name]: {
-        value:
-          field.display && profile[field.name]
-            ? field.display(profile[field.name])
-            : profile[field.name],
-      },
-    }))
-    .reduce((acc, cur) => ({ ...acc, ...cur }), {})
-}
-
-export function removeValidation(profile, rules) {
-  const allRules = [...rules.personalFields, ...rules.businessFields]
-
-  return allRules
-    .map(field => ({
-      [field.name]:
-        field.submit && profile[field.name].value
-          ? field.submit(profile[field.name].value)
-          : profile[field.name].value,
-    }))
-    .reduce((acc, cur) => ({ ...acc, ...cur }), {})
-}
-
 export function applyMask(field, value) {
   return field.mask ? field.mask(value) : value
 }