+++ react/propTypes/index.js
@@ -1,7 +1,0 @@
-export { default as RuleShape } from './RuleShape'
-export { default as RuleFieldShape } from './RuleFieldShape'
-export { default as ProfileShape } from './ProfileShape'
-export { default as ProfileFieldShape } from './ProfileFieldShape'
-export {
-  default as ProfileWithValidationShape,
-} from './ProfileWithValidationShape'