+++ react/rules/default.js
@@ -1,8 +1,3 @@
-import moment from 'moment'
-import msk from 'msk'
-import Phone from '@vtex/phone'
-import brazil from '@vtex/phone/countries/BRA'
-
 export default {
   country: 'BRA',
   personalFields: [
@@ -31,6 +26,11 @@ export default {
       required: true,
     },
     {
+      name: 'homePhone',
+      maxLength: 30,
+      label: 'homePhone',
+    },
+    {
       name: 'gender',
       maxLength: 30,
       label: 'gender',
@@ -39,11 +39,6 @@ export default {
       name: 'birthDate',
       maxLength: 30,
       label: 'birthDate',
-    },
-    {
-      name: 'homePhone',
-      maxLength: 30,
-      label: 'homePhone',
     },
   ],
   businessFields: [